feat: ui 컴포넌트화
