+++ packages/web/src/features/executive/components/SecretKey/types.ts
@@ -1,0 +1,11 @@
+export interface SecretKeyData {
+  id: number;
+  createdAt: Date;
+  secretKey: string;
+  deletedAt: Date | null;
+}
+
+export interface CreateSecretKeyResponse {
+  message: string;
+  createdKey: SecretKeyData;
+}
