+++ packages/web/src/features/executive/frames/ExecutiveDashboardFrame.tsx
@@ -9,9 +9,8 @@ import SectionTitle from "@sparcs-clubs/web/common/components/SectionTitle";
 import Banner from "@sparcs-clubs/web/features/landing/components/Banner";
 
 import DashboardButton from "../components/DashboardButton";
-//import OperationCommitteeSecretManager from "../components/OperationCommitteeSecretManager";
+import OperationCommitteeSecretManager from "../components/SecretKey/OperationCommitteeSecretManager";
 import ManageMemberFrame from "./ManageMemberFrame";
-import ManageSecretKeyFrame from "./ManageSecretKeyFrame";
 import ManageSemesterFrame from "./ManageSemesterFrame";
 
 const DashboardSectionInner = styled.div`
@@ -38,7 +37,7 @@ const ExecutiveDashboardFrame = () => {
       </Banner>
       <ManageSemesterFrame />
       <ManageMemberFrame />
-      <ManageSecretKeyFrame />
+      <OperationCommitteeSecretManager />
 
       <FlexWrapper direction="column" gap={20}>
         <SectionTitle>기간 관리</SectionTitle>
